Introduce daily token usage tracking and limits: add `UsageLog` entity, `UsageLogRepository`, and `UsageService` with methods to record and enforce token usage limits; update `AiGenerationServiceV1Impl` to integrate usage checks and token recording; enhance error handling and OpenAPI documentation configuration.

+++ src/main/java/com/lovable/dto/subscription/UsageTodayResponse.java
@@ -29,8 +29,8 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor
 @Builder
 public class UsageTodayResponse {
-    private int tokensUsed;
-    private int tokensLimit;
-    private int previewsRunning;
-    private int previewsLimit;
+    private Long tokensUsed;
+    private Long tokensLimit;
+    private Long previewsRunning;
+    private Long previewsLimit;
 }
